Revert "Add finalizeBlock(hash)"
