+++ include/veriblock/blockchain/block_index.hpp
@@ -11,9 +11,13 @@
 #include <vector>
 #include <veriblock/arith_uint256.hpp>
 #include <veriblock/blockchain/block_status.hpp>
+#include <veriblock/blockchain/command.hpp>
+#include <veriblock/blockchain/command_group.hpp>
+#include <veriblock/entities/endorsements.hpp>
 #include <veriblock/entities/vbkblock.hpp>
 #include <veriblock/logger.hpp>
 #include <veriblock/validation_state.hpp>
+#include <veriblock/write_stream.hpp>
 
 namespace altintegration {
 
@@ -34,50 +38,6 @@ struct BlockIndex : public Block::addon_t {
 
   //! (memory only) a set of pointers for forward iteration
   std::set<BlockIndex*> pnext{};
-
-  //! (memory only) if true, this block can not be reorganized
-  bool finalized = false;
-
-  BlockIndex(BlockIndex* prev) {
-    if (prev != nullptr) {
-      pprev = prev;
-      prev->pnext.insert(this);
-      height = prev->getHeight() + 1;
-    } else {
-      height = 0;
-    }
-  }
-
-  ~BlockIndex() {
-    // make sure we deleted this block from prev->pnext
-    if (pprev != nullptr) {
-      pprev->pnext.erase(this);
-    }
-
-    // make sure we deleted this block from next blocks pprev.
-    // it is ok to deallocate a block in a middle of chain. we will detect
-    // orphaned blocks and cleanup later.
-    for (auto* next : pnext) {
-      VBK_ASSERT(next);
-      next->pprev = nullptr;
-    }
-  }
-
-  // BlockIndex is not copyable
-  // BlockIndex is movable
-  BlockIndex(BlockIndex&& other) noexcept = default;
-  BlockIndex& operator=(BlockIndex&& other) noexcept = default;
-
-  // returns a copy of BlockIndex without inmem fields
-  BlockIndex<Block> clone() const {
-    BlockIndex<Block> ret = *this;
-    ret.setNullInmemFields();
-    return ret;
-  }
-
-  // loads on-disk fields from 'other' block index.
-  // works like (explicit) copy constructor
-  void mergeFrom(const BlockIndex& other) { *this = other.clone(); }
 
   /**
    * Block is connected if it contains block body (PopData), and all its
@@ -282,15 +242,13 @@ struct BlockIndex : public Block::addon_t {
   }
 
   std::string toPrettyString(size_t level = 0) const {
-    return fmt::sprintf(
-        "%s%sBlockIndex(height=%d, hash=%s, status=%d, header=%s, %s)",
-        std::string(level, ' '),
-        Block::name(),
-        height,
-        HexStr(getHash()),
-        status,
-        header->toPrettyString(),
-        addon_t::toPrettyString());
+    return fmt::sprintf("%s%sBlockIndex(height=%d, hash=%s, status=%d, %s)",
+                        std::string(level, ' '),
+                        Block::name(),
+                        height,
+                        HexStr(getHash()),
+                        status,
+                        addon_t::toPrettyString());
   }
 
   std::string toShortPrettyString() const {
@@ -329,11 +287,6 @@ struct BlockIndex : public Block::addon_t {
   friend bool DeserializeFromVbkEncoding(ReadStream& stream,
                                          BlockIndex<T>& out,
                                          ValidationState& state);
-
- private:
-  // make it non-copyable
-  BlockIndex(const BlockIndex& other) = default;
-  BlockIndex& operator=(const BlockIndex& other) = default;
 };
 
 /**
@@ -343,8 +296,7 @@ struct BlockIndex : public Block::addon_t {
  * the complexity is O(n)
  */
 template <typename Block>
-BlockIndex<Block>* getForkBlock(const BlockIndex<Block>& a,
-                                const BlockIndex<Block>& b) {
+BlockIndex<Block>& getForkBlock(BlockIndex<Block>& a, BlockIndex<Block>& b) {
   const auto initialHeight = std::min(a.getHeight(), b.getHeight());
 
   for (auto cursorA = a.getAncestor(initialHeight),
@@ -352,51 +304,14 @@ BlockIndex<Block>* getForkBlock(const BlockIndex<Block>& a,
        cursorA != nullptr && cursorB != nullptr;
        cursorA = cursorA->getPrev(), cursorB = cursorB->getPrev()) {
     if (cursorA == cursorB) {
-      return cursorA;
+      return *cursorA;
     }
   }
 
-  // chain `b` is not connected to `a`
-  return nullptr;
-}
-
-//! a `candidate` is considered outdated iff it is behind `finalBlock`, or on
-//! same height and not equal to `finalBlock`, or its fork block is outdated
-template <typename Block>
-bool isBlockOutdated(const BlockIndex<Block>& finalBlock,
-                     const BlockIndex<Block>& candidate,
-                     int window = 0) {
-  // finalBlock is ancestor of candidate
-  if (candidate.getAncestor(finalBlock.getHeight()) == &finalBlock) {
-    return false;
-  }
-
-  // all candidates behind final block are outdated
-  if (candidate.getHeight() < finalBlock.getHeight() - window) {
-    return true;
-  }
-
-  // all parallel blocks (on same height as final, but not final) are outdated
-  if (candidate.getHeight() == finalBlock.getHeight() &&
-      &finalBlock != &candidate) {
-    return true;
-  }
-
-  // candidate is ancestor of finalBlock and within window
-  if (finalBlock.getAncestor(candidate.getHeight()) == &candidate) {
-    return false;
-  }
-
-  // candidate is on a fork
-  auto* fork = getForkBlock(finalBlock, candidate);
-
-  // candidate does not connect to finalBlock
-  if (fork == nullptr) {
-    return true;
-  }
-
-  // if fork block is outdated, then candidate is also outdated
-  return isBlockOutdated(finalBlock, *fork);
+  VBK_ASSERT_MSG(false,
+                 "blocks %s and %s must be part of the same tree",
+                 a.toPrettyString(),
+                 b.toPrettyString());
 }
 
 template <typename Block>
