+++ include/veriblock/blockchain/pop/fork_resolution.hpp
@@ -303,7 +303,7 @@ int comparePopScoreImpl(PublicationView& a, PublicationView& b) {
   //
   // Chain A has keystone 60 but chain B does not, so Chain B is
   // chopped to VBK20:BTC100, VBK40:BTC101.
-
+  
   // Also, the chain goes out of finality if it contains a keystone which violates
   // the Bitcoin finality delay based on the previous keystone in the chain.
   // clang-format on
@@ -514,11 +514,6 @@ struct PopAwareForkResolutionComparator {
       return 1;
     }
 
-    if (bestTip->finalized && candidate.getHeight() <= bestTip->getHeight()) {
-      // finalized blocks can not be reorganized
-      return 1;
-    }
-
     if (currentBest.contains(&candidate)) {
       VBK_LOG_INFO(
           "Candidate %s is part of the active chain, the current chain wins",
@@ -562,17 +557,6 @@ struct PopAwareForkResolutionComparator {
         bestTip->toPrettyString(),
         candidate.toPrettyString());
 
-    // if block next to fork is finalized, we don't do FR, since it can not be
-    // reorganized.
-    auto* nextToFork = currentBest[fork->getHeight() + 1];
-    if (nextToFork != nullptr && nextToFork->finalized) {
-      // block `fork+1` is on active chain, and ancestor of currentBest.
-      // which means that `fork+1` can not be reorganized and candidate will
-      // never win.
-      return 1;
-    }
-
-    // multi-keystone POP FR starts
     bool AcrossedKeystoneBoundary =
         isCrossedKeystoneBoundary(fork->getHeight(), bestTip->getHeight(), ki);
     bool BcrossedKeystoneBoundary =
