+++ include/veriblock/blockchain/pop/pop_state_machine.hpp
@@ -306,13 +306,12 @@ struct PopStateMachine {
       return true;
     }
 
-    auto* forkBlock = getForkBlock(from, to);
-    VBK_ASSERT_MSG(forkBlock, "Fork Block must exist!");
+    auto& forkBlock = getForkBlock(from, to);
 
-    unapply(from, *forkBlock);
-    if (!apply(*forkBlock, to, state)) {
+    unapply(from, forkBlock);
+    if (!apply(forkBlock, to, state)) {
       // attempted to switch to an invalid block, rollback
-      bool success = apply(*forkBlock, from, state);
+      bool success = apply(forkBlock, from, state);
       VBK_ASSERT_MSG(success,
                      "state corruption: failed to rollback the state: %s",
                      state.toString());
