+++ include/veriblock/blockchain/pop/vbk_block_tree.hpp
@@ -106,7 +106,7 @@ struct VbkBlockTree : public BlockTree<VbkBlock, VbkChainParams> {
   //! - does validation of endorsements
   //! - recovers tips array
   //! @invariant NOT atomic.
-  bool loadBlock(std::unique_ptr<index_t> index, ValidationState& state) override;
+  bool loadBlock(const index_t& index, ValidationState& state) override;
 
   BtcTree& btc() { return cmp_.getProtectingBlockTree(); }
   const BtcTree& btc() const { return cmp_.getProtectingBlockTree(); }
