+++ include/veriblock/blockchain/vbk_chain_params.hpp
@@ -58,13 +58,6 @@ struct VbkChainParams {
     return 400;
   }
 
-  //! when finalizeBlockImpl is called, this many blocks behind final block will be
-  //! preserved in RAM.
-  //! In VBK this number depends on endorsement settlement interval.
-  uint32_t preserveBlocksBehindFinal() const noexcept {
-    return getEndorsementSettlementInterval();
-  }
-
  protected:
   uint32_t mMaxFutureBlockTime = 5 * 60;  // 5 min
   std::vector<uint32_t> forkResolutionLookUpTable_{
