+++ include/veriblock/storage/inmem_block_storage.hpp
@@ -33,15 +33,14 @@ struct InmemBlockProvider {
   }
 
   template <typename T>
-  std::vector<std::unique_ptr<BlockIndex<T>>> load() {
-    std::vector<std::unique_ptr<BlockIndex<T>>> ret;
+  std::vector<BlockIndex<T>> load() {
+    std::vector<BlockIndex<T>> ret;
     auto& m = getBlocks<T>();
     ret.reserve(m.size());
 
     for (auto& b : m) {
       auto& bi = b.second;
-      auto index = make_unique<BlockIndex<T>>(bi->clone());
-      ret.push_back(std::move(index));
+      ret.push_back(*bi);
     }
 
     return ret;
@@ -72,17 +71,17 @@ struct InmemBlockBatch : public BlockBatch {
 
   void writeBlock(const BlockIndex<BtcBlock>& value) override {
     storage_.btc[value.getHash()] =
-        std::make_shared<BlockIndex<BtcBlock>>(value.clone());
+        std::make_shared<BlockIndex<BtcBlock>>(value);
   }
 
   void writeBlock(const BlockIndex<VbkBlock>& value) override {
     storage_.vbk[value.getHash()] =
-        std::make_shared<BlockIndex<VbkBlock>>(value.clone());
+        std::make_shared<BlockIndex<VbkBlock>>(value);
   }
 
   void writeBlock(const BlockIndex<AltBlock>& value) override {
     storage_.alt[value.getHash()] =
-        std::make_shared<BlockIndex<AltBlock>>(value.clone());
+        std::make_shared<BlockIndex<AltBlock>>(value);
   }
 
   void writeTip(const BlockIndex<BtcBlock>& value) override {
