+++ src/blockchain/alt_block_tree.cpp
@@ -521,21 +521,20 @@ void AltBlockTree::setTipContinueOnInvalid(AltBlockTree::index_t& to) {
   overrideTip(to);
 }
 
-bool AltBlockTree::loadBlock(std::unique_ptr<index_t> index,
+bool AltBlockTree::loadBlock(const AltBlockTree::index_t& index,
                              ValidationState& state) {
-  const auto containingHash = index->getHash();
-  const auto height = index->getHeight();
-  if (!base::loadBlock(std::move(index), state)) {
+  if (!base::loadBlock(index, state)) {
     return false;  // already set
   }
 
   // load endorsements
+  auto containingHash = index.getHash();
   auto* current = getBlockIndex(containingHash);
   VBK_ASSERT(current);
 
-  const auto& vbkblocks = current->getPayloadIds<VbkBlock>();
-  const auto& vtbs = current->getPayloadIds<VTB>();
-  const auto& atvs = current->getPayloadIds<ATV>();
+  auto vbkblocks = current->getPayloadIds<VbkBlock>();
+  auto vtbs = current->getPayloadIds<VTB>();
+  auto atvs = current->getPayloadIds<ATV>();
   if (hasDuplicates<VbkBlock>(*current, vbkblocks, *this, state) ||
       hasDuplicates<VTB>(*current, vtbs, *this, state) ||
       hasDuplicates<ATV>(*current, atvs, *this, state)) {
@@ -543,8 +542,8 @@ bool AltBlockTree::loadBlock(std::unique_ptr<index_t> index,
   }
 
   // recover `endorsedBy` and `blockOfProofEndorsements`
-  const int si = getParams().getEndorsementSettlementInterval();
-  auto window = std::max(0, height - si);
+  auto window = std::max(
+      0, index.getHeight() - getParams().getEndorsementSettlementInterval());
   Chain<index_t> chain(window, current);
   if (!recoverEndorsements(*this, chain, *current, state)) {
     return state.Invalid("bad-endorsements");
@@ -616,11 +615,6 @@ std::vector<const AltBlockTree::index_t*> AltBlockTree::getConnectedTipsAfter(
   }
 
   return candidates;
-}
-
-bool AltBlockTree::finalizeBlock(const AltBlockTree::hash_t& block) {
-  return base::finalizeBlockImpl(block,
-                                 getParams().preserveBlocksBehindFinal());
 }
 
 template <typename Payloads>
