+++ test/blockchain/CMakeLists.txt
@@ -36,9 +36,13 @@ addtest(btc_blockchain_test
         ${btc_gen_1}
         ${btc_gen_2}
         )
+set_target_properties(btc_blockchain_test PROPERTIES
+        COST 10000  # 10 sec
+        )
 
 addtest(vbk_blockchain_test
         vbk_blockchain_test.cpp
+        vbk_blockchain_util_test.cpp
         # generated cpps
         ${vbk_gen_1}
         ${vbk_gen_2}
@@ -55,14 +59,18 @@ set_tests_properties(vbk_getblockproof_test PROPERTIES
         COST 30000 # 30 sec
         )
 
-addtest(chain_test chain_test.cpp)
-addtest(blockchain_test blockchain_test.cpp)
-set_test_cost(blockchain_test 105000) # 105 sec
+addtest(blockchain_test
+        chain_test.cpp
+        blockchain_test.cpp
+        )
+set_tests_properties(blockchain_test PROPERTIES
+        COST 10000  # 10 sec
+        )
+
+
 addtest(chainparams_test chainparams_test.cpp)
 addtest(alt_blockchain_test alt_blockchain_test.cpp)
 addtest(alt_invalidation_test alt_invalidation_test.cpp)
 addtest(btc_invalidation_test btc_invalidation_test.cpp)
-addtest(vbk_blockchain_util_test vbk_blockchain_util_test.cpp)
 addtest(temp_block_tree_test temp_block_tree_test.cpp)
 addtest(alt_tree_bootstrap_test alt_tree_bootstrap_test.cpp)
-addtest(block_finalization_test block_finalization_test.cpp)
