+++ test/blockchain/vbk_blockchain_util_test.cpp
@@ -17,23 +17,23 @@
 
 using namespace altintegration;
 
-static std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> getChain(
-    int32_t deltaTime,
-    int32_t deltaTime_change,
-    uint32_t difficulty,
-    uint32_t chainlength) {
+static std::vector<BlockIndex<VbkBlock>> getChain(int32_t deltaTime,
+                                                  int32_t deltaTime_change,
+                                                  uint32_t difficulty,
+                                                  uint32_t chainlength) {
   assert(chainlength != 0);
 
   VbkBlock block{};
   block.setHeight(1);
   block.setTimestamp(10000);
   block.setDifficulty(difficulty);
-  auto blockIndex = std::make_shared<BlockIndex<VbkBlock>>(nullptr);
-  blockIndex->setHeader(block);
-  blockIndex->setHeight(block.getHeight());
+  BlockIndex<VbkBlock> blockIndex;
+  blockIndex.setHeader(block);
+  blockIndex.setHeight(block.getHeight());
+  blockIndex.pprev = nullptr;
 
-  std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> chain(chainlength);
-  chain[0] = std::move(blockIndex);
+  std::vector<BlockIndex<VbkBlock>> chain(chainlength);
+  chain[0] = blockIndex;
 
   for (size_t i = 1; i < chainlength; ++i) {
     if (i > 0 && i % 40 == 0) {
@@ -42,13 +42,14 @@ static std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> getChain(
 
     VbkBlock blockTmp{};
     blockTmp.setHeight((int32_t)i + 1);
-    blockTmp.setTimestamp(chain[i - 1]->getHeader().getBlockTime() + deltaTime);
-    blockTmp.setDifficulty(difficulty);
-    auto temp = std::make_shared<BlockIndex<VbkBlock>>(chain[i - 1].get());
-    temp->setHeader(blockTmp);
-    temp->setHeight(blockTmp.getHeight());
+    blockTmp.setTimestamp(chain[i - 1].getHeader().getBlockTime() + deltaTime);
+    blockTmp.setDifficulty( difficulty);
+    BlockIndex<VbkBlock> temp;
+    temp.setHeader(blockTmp);
+    temp.setHeight(blockTmp.getHeight());
+    temp.pprev = &chain[i - 1];
 
-    chain[i] = std::move(temp);
+    chain[i] = temp;
   }
 
   return chain;
@@ -151,13 +152,13 @@ static std::vector<GetNextWorkRequiredTestCases>
 TEST_P(GetNextWorkRequiredTest, getNextWorkRequired_test) {
   auto value = GetParam();
 
-  auto chain = getChain(value.deltaTime,
-                        value.deltaTime_change,
-                        value.chain_difficulty,
-                        value.chainlength);
+  std::vector<BlockIndex<VbkBlock>> chain = getChain(value.deltaTime,
+                                                     value.deltaTime_change,
+                                                     value.chain_difficulty,
+                                                     value.chainlength);
 
   uint32_t result =
-      getNextWorkRequired(*chain[chain.size() - 1], VbkBlock(), *chainparams);
+      getNextWorkRequired(chain[chain.size() - 1], VbkBlock(), *chainparams);
 
   EXPECT_EQ(value.expected_difficulty, result);
 }
@@ -176,13 +177,13 @@ TEST_F(SingleTest, single_test) {
   block.setHeight(1);
   block.setTimestamp(10000);
   block.setDifficulty(ArithUint256::fromHex("09184E72A000").toBits());
-  auto blockIndex = std::make_shared<BlockIndex<VbkBlock>>(nullptr);
-  blockIndex->setHeader(block);
-  blockIndex->setHeight(block.getHeight());
-  blockIndex->pprev = nullptr;
+  BlockIndex<VbkBlock> blockIndex;
+  blockIndex.setHeader(block);
+  blockIndex.setHeight(block.getHeight());
+  blockIndex.pprev = nullptr;
 
-  std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> chain(chainlength);
-  chain[0] = std::move(blockIndex);
+  std::vector<BlockIndex<VbkBlock>> chain(chainlength);
+  chain[0] = blockIndex;
 
   for (size_t i = 1; i < chainlength; ++i) {
     if (i >= 60) {
@@ -191,17 +192,18 @@ TEST_F(SingleTest, single_test) {
 
     VbkBlock blockTmp{};
     blockTmp.setHeight((int32_t)i + 1);
-    blockTmp.setTimestamp(chain[i - 1]->getHeader().getBlockTime() + deltaTime);
+    blockTmp.setTimestamp(chain[i - 1].getHeader().getBlockTime() + deltaTime);
     blockTmp.setDifficulty(ArithUint256::fromHex("09184E72A000").toBits());
-    auto temp = std::make_shared<BlockIndex<VbkBlock>>(chain[i - 1].get());
-    temp->setHeader(blockTmp);
-    temp->setHeight(blockTmp.getHeight());
+    BlockIndex<VbkBlock> temp;
+    temp.setHeader(blockTmp);
+    temp.setHeight(blockTmp.getHeight());
+    temp.pprev = &chain[i - 1];
 
-    chain[i] = std::move(temp);
+    chain[i] = temp;
   }
 
   uint32_t result =
-      getNextWorkRequired(*chain.back(), VbkBlock(), *chainparams);
+      getNextWorkRequired(chain[chain.size() - 1], VbkBlock(), *chainparams);
 
   EXPECT_EQ(ArithUint256::fromHex("0228C35294D0").toBits(), result);
 }
@@ -211,14 +213,13 @@ TEST(Vbk, CheckBlockTime1) {
   const auto startTime = 1'527'000'000;
   std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> chain;
   for (int i = 0; i < 1000; i++) {
-    auto b = std::make_shared<BlockIndex<VbkBlock>>(
-        i == 0 ? nullptr : chain.back().get());
-    chain.push_back(std::move(b));
-    auto& index = chain.back();
+    chain.push_back(std::make_shared<BlockIndex<VbkBlock>>());
+    auto& index = chain[chain.size() - 1];
     VbkBlock blockTmp{};
     blockTmp.setTimestamp(startTime + (120 * i));
     index->setHeight(VBK_MINIMUM_TIMESTAMP_ONSET_BLOCK_HEIGHT + i);
     index->setHeader(blockTmp);
+    index->pprev = i == 0 ? nullptr : chain[i - 1].get();
   }
 
   VbkChainParamsMain main;
@@ -239,35 +240,38 @@ TEST(Vbk, CheckBlockTime1) {
 }
 
 TEST(Vbk, CheckBlockTime2) {
-  auto makeBlock = [](BlockIndex<VbkBlock>* prev,
-                      int timestamp,
-                      int height) -> std::shared_ptr<BlockIndex<VbkBlock>> {
-    auto index = std::make_shared<BlockIndex<VbkBlock>>(prev);
+  auto makeBlock = [](int timestamp, int height) -> BlockIndex<VbkBlock> {
+    BlockIndex<VbkBlock> index;
     VbkBlock blockTmp{};
     blockTmp.setTimestamp(timestamp);
-    index->setHeight(height);
-    index->setHeader(blockTmp);
+    index.setHeight(height);
+    index.setHeader(blockTmp);
     return index;
   };
 
   ValidationState state;
-  std::vector<std::shared_ptr<BlockIndex<VbkBlock>>> chain;
-  chain.push_back(makeBlock(nullptr, 1527000000, 110000));
-  chain.push_back(makeBlock(chain.back().get(), 1527500000, 110001));
-  chain.push_back(makeBlock(chain.back().get(), 1528000000, 110002));
+  std::vector<BlockIndex<VbkBlock>> chain;
+  chain.push_back(makeBlock(1527000000, 110000));
+  chain.push_back(makeBlock(1527500000, 110001));
+  chain.push_back(makeBlock(1528000000, 110002));
+
+  for (size_t i = 0; i < chain.size(); i++) {
+    auto& index = chain[i];
+    index.pprev = i == 0 ? nullptr : &chain[i - 1];
+  }
 
   VbkBlock block;
   VbkChainParamsMain params;
   block.setTimestamp(1527499999);
   bool r1 = checkBlockTime<VbkBlock, VbkChainParams>(
-      *chain[chain.size() - 1], block, state, params);
+      chain[chain.size() - 1], block, state, params);
   ASSERT_FALSE(r1);
   ASSERT_EQ(state.GetPathParts()[state.GetPathParts().size() - 1],
             "vbk-time-too-old");
 
   block.setTimestamp(1527500000);
   bool r2 = checkBlockTime<VbkBlock, VbkChainParams>(
-      *chain[chain.size() - 1], block, state, params);
+      chain[chain.size() - 1], block, state, params);
   ASSERT_TRUE(r2);
 }
