+++ test/crypto/progpow/CMakeLists.txt
@@ -8,9 +8,14 @@ addtest(progpow_test
         progpow_test.cpp
         ${progpow_test_vectors}
         )
-set_test_cost(progpow_test 38000) # 38 sec
+set_tests_properties(progpow_test PROPERTIES
+        COST 9000 # ~9 sec
+        )
 addtest(ethash_dag_test ethash_dag_test.cpp)
-set_test_cost(ethash_dag_test 21000) # 21 sec
-
+set_tests_properties(ethash_dag_test PROPERTIES
+        COST 26000 # ~26 sec
+        )
 addtest(ethash_dagseed_test ethash_dagseed_test)
-set_test_cost(ethash_dagseed_test 10000) # 10 sec
+set_tests_properties(ethash_dagseed_test PROPERTIES
+        COST 10000 # ~10 sec
+        )
