+++ test/rewards/CMakeLists.txt
@@ -1,2 +1,2 @@
-addtest(poprewards_test poprewards_test.cpp)
-set_test_cost(poprewards_test 33000) # 33 sec
+addtest(poprewards_test
+        poprewards_test.cpp)
