+++ test/storage/CMakeLists.txt
@@ -1,3 +1,2 @@
 addtest(alttree_storage_test alttree_storage_test.cpp)
 addtest(save_load_tree_test save_load_tree_test.cpp)
-set_test_cost(save_load_tree_test 22000) # 22 sec
