+++ test/storage/save_load_tree_test.cpp
@@ -36,8 +36,6 @@ struct SaveLoadTreeTest : public PopTestFixture, public testing::Test {
   }
 
   auto assertTreesEqual() {
-    assertTreesHaveNoOrphans(alttree);
-    assertTreesHaveNoOrphans(alttree2);
     ASSERT_TRUE(cmp(alttree, alttree2));
     auto tip = alttree.getBestChain().tip();
     auto to1 = alttree.getBestChain().first()->getHash();
